collapse to single file
