+++ src/console/mod.ts
@@ -3,8 +3,8 @@ export type { ConfirmOptions } from "./confirm.ts";
 export { logger } from "./logger.ts";
 export { maybeMultiSelect, multiSelect } from "./multiSelect.ts";
 export type { MultiSelectOption, MultiSelectOptions } from "./multiSelect.ts";
-export type { ProgressOptions } from "./progress/mod.ts";
-export { isShowingProgressBars, ProgressBar } from "./progress/mod.ts";
+export type { ProgressOptions } from "./progress.ts";
+export { isShowingProgressBars, ProgressBar } from "./progress.ts";
 export { maybePrompt, prompt } from "./prompt.ts";
 export type { PromptInputMask, PromptOptions } from "./prompt.ts";
 export { maybeSelect, select } from "./select.ts";
